requirements.txt
pandas>=2.1.4
numpy>=1.26.0
opencv-python>=4.8.1.78
Pillow>=10.1.0
pytesseract>=0.3.10
PyPDF2>=3.0.1
PyMuPDF>=1.23.0
python-docx>=1.1.0
openpyxl>=3.1.2
python-pptx>=0.6.23
presidio-analyzer>=2.2.33
presidio-anonymizer>=2.2.33
scikit-learn>=1.3.2
nltk>=3.8.1
regex>=2023.10.3
jupyter>=1.0.0
matplotlib>=3.8.2
seaborn>=0.13.0
plotly>=5.17.0
streamlit>=1.28.2
